{"schema":"decodex/commit/1","summary":"Split consolidation proposal review event persistence into a private sibling module after local validation.","authority":"manual"}
